Crate buerostatus

This crate is a small wrapper around the buerostatus API of the @fsr that tells you whether someone is in our office or not.

Example

if let Ok(is_open) = get_buerostatus() {
    if is_open { println!("Someone's inside!"); }
    else { println!("No one is there..."); }
}
else {
    println!("An error occured!");
}

The function get_buerostatus() offers a precise error message in case of an error. See the enum ApiError for more information.

Enums

ApiError
A set of errors that may occur during the API call.

Functions

get_buerostatus
Gets the buerostatus from ifsr.de.