Function buerostatus::get_buerostatus

pub fn get_buerostatus() -> Result<bool, ApiError>

Gets the buerostatus from ifsr.de.

Returns true, when the office is open, false if not and an ApiError if any form of error is occured during execution.

Example

if let Ok(is_open) = get_buerostatus() {
    if is_open {
        println!("Someone's inside!");
    }
}
else {
    // Error Handling...
}